refactor(emails): migrate method `send_email` to utils module
Migrate method to utils module. It is the only method in the
emails module, so the utils module is not a worde place for it.
Also moving the method and getting rid of the module reduces
complexity of the codebase.
